1.8.19    Improvements and bug fix.

+++ src/main/java/com/landawn/abacus/util/stream/ArrayByteStream.java
@@ -1241,7 +1241,7 @@ class ArrayByteStream extends AbstractByteStream {
     }
 
     @Override
-    public int sum() {
+    public long sum() {
         assertNotClosed();
 
         try {
