+++ src/main/java/com/landawn/abacus/util/stream/ArrayCharStream.java
@@ -1241,7 +1241,7 @@ class ArrayCharStream extends AbstractCharStream {
     }
 
     @Override
-    public int sum() {
+    public long sum() {
         assertNotClosed();
 
         try {
